Added PXI base init to default init routine
libctr_core, in order to support the different PXI addresses used by
both processes, now requires for the base address to be set explicitly.
For convenience, this is now set automatically during init to the
physical location in memory. If virtual memory is set up, the address
needs to be updated if the IO registers are mapped to another location.