Screen/surface work, cache, MMU, crt0, autotools
-Converted screen/surface stuff to use libctr_core's new C++ way of
handling these things.
-Tried to implement caching routines, haven't tested them. Some are
missing, namely branch target cache management.
-Implemented MMU support. Thanks to @TuxSH and @profi200 for poking me
in the general direction of getting this working. MMU is initialized
with a flat map, with the IO region set to be Shared Device memory, and
everything else normal cached.
-As part of the MMU changes, unaligned instruction support has been
enabled in the crt0. This is due to the fact that GCC, by default, for
 -mcpu=mpcore assumes unaligned accesses are possible.
-Moved around a lot of crt0's functionality, to make sure the stack and
.bss are initialized before messing with MMU C functions.
-Updated copyright headers.
-Overhauled autoconf and automake scripts. The previous implementation
was full of holes. This one is as well, probably, but more warnings are
checked, C*FLAGS handling is done better, and C*FLAGS are set before
LT_INIT is invoked to make sure libtool finds the right libraries it
wants (namely libstdc++).